mod: refactored and split model with trainer
